# Trivial optimize on ifNotNull with intel AVX2 