### Changelog category (leave one):
- Performance Improvement



### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Micro benchmarks: https://github.com/bigo-sg/gluten/commit/b1fbaaeb6d1e3edd83c4158993c31e6a2adc5d3d 
``` 
Run on (32 X 2100 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x16)
  L1 Instruction 32 KiB (x16)
  L2 Unified 1024 KiB (x16)
  L3 Unified 11264 KiB (x2)
Load Average: 4.47, 4.43, 4.74
-------------------------------------------------------------------
Benchmark                         Time             CPU   Iterations
-------------------------------------------------------------------
BM_isNotNull                   3884 ns         3884 ns       180237
BM_isNotNullSSE42              3893 ns         3893 ns       179702
BM_isNotNullAVX2               2983 ns         2982 ns       236436
BM_isNotNullAVX512F            2856 ns         2855 ns       239358
BM_isNotNullAVX512BW           2968 ns         2968 ns       238690
``` 

Use AVX2 for isNotNull is fast enough to avoid too much code.  Let's see if performance test in CI shows any benefit